+++ HackIllinois/Controllers/ViewControllers/HIAnnouncmentsViewController.swift
@@ -10,7 +10,7 @@ import Foundation
 import UIKit
 import CoreData
 
-class HIAnnouncementsViewController: HIBaseViewController {
+class HIAnnouncmentsViewController: HIBaseViewController {
     // MARK - Properties
     lazy var fetchedResultsController: NSFetchedResultsController<Announcement> = {
         let fetchRequest: NSFetchRequest<Announcement> = Announcement.fetchRequest()
@@ -23,31 +23,15 @@ class HIAnnouncementsViewController: HIBaseViewController {
 
         return fetchedResultsController
     }()
-
 }
 
 // MARK: - Actions
-extension HIAnnouncementsViewController {
+extension HIAnnouncmentsViewController {
 
 }
 
 // MARK: - UIViewController
-extension HIAnnouncementsViewController {
-    override func loadView() {
-        super.loadView()
-
-        let tableView = UITableView(frame: .zero, style: .grouped)
-        tableView.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(tableView)
-        
-        tableView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor).isActive = true
-        tableView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor).isActive = true
-        tableView.bottomAnchor.constraint(equalTo: view.bottomAnchor).isActive = true
-        tableView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor).isActive = true
-        
-        self.tableView = tableView
-    }
-    
+extension HIAnnouncmentsViewController {
     override func viewDidLoad() {
         tableView?.register(UINib(nibName: HIAnnouncementCell.IDENTIFIER, bundle: nil), forCellReuseIdentifier: HIAnnouncementCell.IDENTIFIER)
         super.viewDidLoad()
@@ -56,12 +40,10 @@ extension HIAnnouncementsViewController {
         
         try? _fetchedResultsController?.performFetch()
     }
-    
-
 }
 
 // MARK: - UINavigationItem Setup
-extension HIAnnouncementsViewController {
+extension HIAnnouncmentsViewController {
     @objc dynamic override func setupNavigationItem() {
         super.setupNavigationItem()
         title = "NOTIFICATIONS"
@@ -69,7 +51,7 @@ extension HIAnnouncementsViewController {
 }
 
 // MARK: - UITableViewDataSource
-extension HIAnnouncementsViewController {
+extension HIAnnouncmentsViewController {
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: HIAnnouncementCell.IDENTIFIER, for: indexPath)
         if let cell = cell as? HIAnnouncementCell {
