+++ HackIllinois/Controllers/ViewControllers/HIEventViewControllers/HIHomeViewController.swift
@@ -27,57 +27,6 @@ class HIHomeViewController: HIEventListViewController {
 
 // MARK: - UIViewController
 extension HIHomeViewController {
-    override func loadView() {
-        super.loadView()
-        
-        let hackingEndsInLabel = UILabel()
-        hackingEndsInLabel.text = "HACKING ENDS IN:"
-        hackingEndsInLabel.textColor = HIColor.darkIndigo
-        hackingEndsInLabel.font = UIFont.systemFont(ofSize: 15, weight: .medium)
-        hackingEndsInLabel.textAlignment = .center
-        hackingEndsInLabel.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(hackingEndsInLabel)
-        hackingEndsInLabel.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 20).isActive = true
-        hackingEndsInLabel.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor).isActive = true
-        hackingEndsInLabel.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor).isActive = true
-        
-        
-        let countdownContainerView = UIView()
-        // TODO: remove background color after view debugging
-        countdownContainerView.backgroundColor = HIColor.hotPink
-        countdownContainerView.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(countdownContainerView)
-        countdownContainerView.topAnchor.constraint(equalTo: hackingEndsInLabel.bottomAnchor, constant: 15).isActive = true
-        countdownContainerView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 12).isActive = true
-        countdownContainerView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor, constant: -12).isActive = true
-        countdownContainerView.heightAnchor.constraint(equalToConstant: 188).isActive = true
-        
-        
-        let happeningNowLabel = UILabel()
-        // TODO: check if colon should appear after NOW
-        happeningNowLabel.text = "HAPPENING NOW"
-        happeningNowLabel.textColor = HIColor.darkIndigo
-        happeningNowLabel.font = UIFont.systemFont(ofSize: 15, weight: .medium)
-        happeningNowLabel.textAlignment = .center
-        happeningNowLabel.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(happeningNowLabel)
-        happeningNowLabel.topAnchor.constraint(equalTo: countdownContainerView.bottomAnchor, constant: 15).isActive = true
-        happeningNowLabel.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor).isActive = true
-        happeningNowLabel.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor).isActive = true
-        
-        
-        let tableView = UITableView(frame: .zero, style: .grouped)
-        tableView.insetsContentViewsToSafeArea = false
-        tableView.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(tableView)
-        tableView.topAnchor.constraint(equalTo: happeningNowLabel.bottomAnchor, constant: 5).isActive = true
-        tableView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor).isActive = true
-        tableView.bottomAnchor.constraint(equalTo: view.bottomAnchor).isActive = true
-        tableView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor).isActive = true
-        
-        self.tableView = tableView
-    }
-    
     override func viewDidLoad() {
         _fetchedResultsController = fetchedResultsController as? NSFetchedResultsController<NSManagedObject>
         super.viewDidLoad()
@@ -89,23 +38,5 @@ extension HIHomeViewController {
     @objc dynamic override func setupNavigationItem() {
         super.setupNavigationItem()
         title = "HOME"
-    }
-}
-
-// MARK: - UITableViewDataSource
-extension HIHomeViewController {
-    override func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
-        return nil
-    }
-}
-
-// MARK: - UITableViewDelegate
-extension HIHomeViewController {
-    override func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-        return .leastNonzeroMagnitude
-    }
-
-    override func tableView(_ tableView: UITableView, heightForFooterInSection section: Int) -> CGFloat {
-        return .leastNonzeroMagnitude
     }
 }
