+++ HackIllinois/Controllers/ViewControllers/HIEventViewControllers/HIScheduleViewController.swift
@@ -59,30 +59,6 @@ extension HIScheduleViewController {
 
 // MARK: - UIViewController
 extension HIScheduleViewController {
-    override func loadView() {
-        super.loadView()
-        
-        let segmentedControl = HISegmentedControl()
-        segmentedControl.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(segmentedControl)
-        
-        segmentedControl.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor).isActive = true
-        segmentedControl.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 16).isActive = true
-        segmentedControl.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor, constant: -16).isActive = true
-        segmentedControl.heightAnchor.constraint(equalToConstant: 34).isActive = true
-        
-        let tableView = UITableView(frame: .zero, style: .grouped)
-        tableView.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(tableView)
-        
-        tableView.topAnchor.constraint(equalTo: segmentedControl.bottomAnchor).isActive = true
-        tableView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor).isActive = true
-        tableView.bottomAnchor.constraint(equalTo: view.bottomAnchor).isActive = true
-        tableView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor).isActive = true
-        
-        self.tableView = tableView
-    }
-    
     override func viewDidLoad() {
         _fetchedResultsController = fetchedResultsController as? NSFetchedResultsController<NSManagedObject>
         super.viewDidLoad()
@@ -96,19 +72,3 @@ extension HIScheduleViewController {
         title = "SCHEDULE"
     }
 }
-
-// MARK: - UITableViewDelegate
-extension HIEventListViewController {
-    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-        return 25
-    }
-
-    func tableView(_ tableView: UITableView, heightForFooterInSection section: Int) -> CGFloat {
-        return 15
-    }
-}
-
-
-
-
-
