--- HackIllinois/Views/HIDateHeader.swift
@@ -1,48 +1,0 @@
-//
-//  HIDateHeader.swift
-//  HackIllinois
-//
-//  Created by Rauhul Varma on 11/22/17.
-//  Copyright © 2017 HackIllinois. All rights reserved.
-//
-
-import Foundation
-import UIKit
-
-class HIDateHeader: UITableViewHeaderFooterView {
-    static let IDENTIFIER = "HIDateHeader"
-
-    var titleLabel = UILabel()
-
-    override init(reuseIdentifier: String?) {
-        super.init(reuseIdentifier: reuseIdentifier)
-
-        let backgroundView = UIView()
-        backgroundView.backgroundColor = HIColor.paleBlue
-        backgroundView.translatesAutoresizingMaskIntoConstraints = false
-        addSubview(backgroundView)
-        backgroundView.topAnchor.constraint(equalTo: topAnchor).isActive = true
-        backgroundView.leadingAnchor.constraint(equalTo: leadingAnchor).isActive = true
-        backgroundView.bottomAnchor.constraint(equalTo: bottomAnchor).isActive = true
-        backgroundView.trailingAnchor.constraint(equalTo: trailingAnchor).isActive = true
-
-        self.backgroundView = backgroundView
-
-        titleLabel.textColor = HIColor.darkIndigo
-        titleLabel.font = UIFont.systemFont(ofSize: 13, weight: .bold)
-        titleLabel.translatesAutoresizingMaskIntoConstraints = false
-        backgroundView.addSubview(titleLabel)
-        titleLabel.leadingAnchor.constraint(equalTo: backgroundView.safeAreaLayoutGuide.leadingAnchor, constant: 14).isActive = true
-        titleLabel.bottomAnchor.constraint(equalTo: backgroundView.safeAreaLayoutGuide.bottomAnchor).isActive = true
-        titleLabel.trailingAnchor.constraint(equalTo: backgroundView.safeAreaLayoutGuide.trailingAnchor, constant: -14).isActive = true
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) should not be used.")
-    }
-
-    override func prepareForReuse() {
-        super.prepareForReuse()
-        titleLabel.text = nil
-    }
-}
