+++ HackIllinois/Views/HIDateHeader/HIDateHeader.swift
@@ -1,0 +1,21 @@
+//
+//  HIDateHeader.swift
+//  HackIllinois
+//
+//  Created by Rauhul Varma on 11/22/17.
+//  Copyright © 2017 HackIllinois. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+class HIDateHeader: UITableViewHeaderFooterView {
+    static let IDENTIFIER = "HIDateHeader"
+
+    @IBOutlet weak var titleLabel: UILabel!
+
+    override func prepareForReuse() {
+        super.prepareForReuse()
+        titleLabel.text = nil
+    }
+}
