+++ HackIllinois/Views/HIDateHeader/HIDateHeader.xib
@@ -1,0 +1,55 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="13771" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina4_7" orientation="portrait">
+        <adaptation id="fullscreen"/>
+    </device>
+    <dependencies>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="13772"/>
+        <capability name="Named colors" minToolsVersion="9.0"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <view contentMode="scaleToFill" id="28u-PF-Jna" customClass="HIDateHeader" customModule="HackIllinois" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="375" height="25"/>
+            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+            <subviews>
+                <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="Aib-bA-8g8">
+                    <rect key="frame" x="0.0" y="0.0" width="375" height="25"/>
+                    <color key="backgroundColor" name="paleBlue"/>
+                </view>
+                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="1iM-6s-ASW">
+                    <rect key="frame" x="14" y="9" width="347" height="16"/>
+                    <fontDescription key="fontDescription" type="boldSystem" pointSize="13"/>
+                    <color key="textColor" name="darkIndigo"/>
+                    <nil key="highlightedColor"/>
+                </label>
+            </subviews>
+            <constraints>
+                <constraint firstAttribute="trailing" secondItem="Aib-bA-8g8" secondAttribute="trailing" id="0LY-9y-ICi"/>
+                <constraint firstItem="Aib-bA-8g8" firstAttribute="top" secondItem="28u-PF-Jna" secondAttribute="top" id="GZ2-O4-ZFg"/>
+                <constraint firstItem="yU8-9J-dQc" firstAttribute="bottom" secondItem="1iM-6s-ASW" secondAttribute="bottom" id="R5q-oU-qpc"/>
+                <constraint firstItem="yU8-9J-dQc" firstAttribute="trailing" secondItem="1iM-6s-ASW" secondAttribute="trailing" constant="14" id="Xpw-3K-TKC"/>
+                <constraint firstItem="Aib-bA-8g8" firstAttribute="leading" secondItem="yU8-9J-dQc" secondAttribute="leading" id="ZZR-gA-dfm"/>
+                <constraint firstItem="1iM-6s-ASW" firstAttribute="leading" secondItem="yU8-9J-dQc" secondAttribute="leading" constant="14" id="l7L-XA-KCL"/>
+                <constraint firstAttribute="bottom" secondItem="Aib-bA-8g8" secondAttribute="bottom" id="oWG-ad-Zku"/>
+            </constraints>
+            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
+            <viewLayoutGuide key="safeArea" id="yU8-9J-dQc"/>
+            <connections>
+                <outlet property="titleLabel" destination="1iM-6s-ASW" id="6Ks-OT-znQ"/>
+            </connections>
+            <point key="canvasLocation" x="101.5" y="-156"/>
+        </view>
+    </objects>
+    <resources>
+        <namedColor name="darkIndigo">
+            <color red="0.10999999940000001" green="0.019999999550000001" blue="0.36899998779999998" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+        </namedColor>
+        <namedColor name="paleBlue">
+            <color red="0.94901960784313721" green="0.96078431372549022" blue="0.99607843137254903" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+        </namedColor>
+    </resources>
+</document>
